requirements.txt
streamlit==1.32.2
face-recognition==1.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
pillow==10.2.0
cmake==3.28.3
dlib==19.24.2